[02旅游景点智能推荐平台]  重构推荐和评论模块
- 将RankedRecommend、RecallCandidate、RecommendResultItem和FavoriteVO中的rankScore和baseScore类型从BigDecimal更改为Double。

- 添加了新的实体ReviewLike和ReviewReply及其对应的映射器。

- 删除了未使用的AI相关API文件并更新了相关文档。

- 增强了ReviewVO和ScenicDetailVO，添加了额外的字段并改进了注释。

- 更新了SQL模式，包括review_like和review_reply表。

- 添加了新的API端点，用于点赞和回复评论，包括文档中的请求/响应规范。

+++ projects/02travel_recommend_system/travel-model/src/main/java/com/travel/advisor/domain/recommend/RankedRecommend.java
@@ -4,7 +4,6 @@ import com.travel.advisor.entity.ScenicSpot;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 /**
@@ -19,7 +18,7 @@ public class RankedRecommend {
     /**
      * 综合得分，基于多个召回策略的基础分数累加计算得出，代表了该景点的推荐优先级。
      */
-    private BigDecimal rankScore;
+    private Double rankScore;
 
     /**
      * 来源类型集合，表示该景点是通过哪些召回策略命中的，例如“基于标签”、“基于地理位置”、“基于用户行为”等。这个字段可以帮助前端展示推荐理由或者进行后续的分析和优化。
